COMPATIBILITY CHANGE: By default now the daemon starts in "atomic" mode.
Atomic mode allows each worker process to perform a single task, with
the intention to keep a clean environment each time a task is to be
executed. This means that by default each worker sits waiting for tasks.
When a task arrives for that worker, it's executed, and the worker
finishes its process, thus having the daemon spawn a new worker.

To disable this behavior specify --atomic=0 in the command line.